Moved to common package
