+++ src/main/java/com/calzyr/services/common/LoggingService.java
@@ -1,4 +1,4 @@
-package com.calzyr.services;
+package com.calzyr.services.common;
 
 import org.springframework.stereotype.Service;
 
@@ -6,7 +6,6 @@ import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.sql.Timestamp;
 import java.util.Date;
 import java.util.logging.Logger;
